(refactor) rename typo and use asyncio.run
